requirements.txt
Flask==3.0.3
keras
tensorflow
pillow
numpy
